Modify error handling in context of votes
If not in debug mode, Nuvotifier will now only log exceptions if we
haven't already sent out the vote as an event. If we have sent out the
vote as an event, then we instead ignore the log.

When in debug mode, the message is now modified specifying that while
there was an exception, be noted that the vote was indeed processed to
the fullest ability to which it could (aka, not our fault!)