Corrected offline vote caching to be case-insensitive
This patch is to mitigate instances where offline votes cast using case-incorrect spelling fails to forward the vote for a player when they next login. The player cannot vote again, and rewards are lost.

Example scenario:

Player with a name Notch (spelled exactly) votes at some website and spells their name using entirely lower-case such as "notch", because they are using a smartphone. The vote is successfully received by the server, but they are unable to receive their reward once they log in, and cannot vote again.

This issue only presents when vote caching is used, which requires `pluginMessaging` to be specified as the forwarding method in the NuVotifier configuration file for BungeeCord. It does not exhibit with the other `proxy` forwarding method, or when NuVotifier is used without BungeeCord.